return piece cache stats

+++ internal/rpctypes/rpctypes.go
@@ -39,8 +39,8 @@ type SessionStats struct {
 	BlockListRules                int
 	BlockListLastSuccessfulUpdate *Time
 	PieceCacheItems               int
-	PieceCacheSize                int
-	ActivePieceBytes              int
+	PieceCacheSize                int64
+	ActivePieceBytes              int64
 }
 
 type Stats struct {
